Commented out EditorUtility functions and Set Latency Test as active
EditorUtility functions can only be used in editor and cause errors when trying to create a build.

Most of the Editor Utility functions were SetDirty commands which as far as I am aware causes editor elements to be redrawn so these can simply not be run if we are not in editor.

Functionality that needs to be replaced has been marked with //TODO:

However the file loading one may not be necessary are files can just be replayed in editor and we can skip this functionality for builds.